backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.3
python-dotenv==1.0.1
reportlab==4.2.2
Werkzeug==3.0.3